Simplify configs to a single file
