--- src/configs/callbacks.py
@@ -1,18 +1,0 @@
-import os
-
-from hydra_zen import builds
-from lightning.pytorch.callbacks import EarlyStopping, ModelCheckpoint, RichModelSummary, RichProgressBar
-
-from src.configs.paths import output_dir
-
-RichProgressBarCfg = builds(RichProgressBar)
-RichModelSummaryCfg = builds(RichModelSummary, max_depth=-1)
-EarlyStoppingCfg = builds(EarlyStopping, monitor="${monitor}", patience=3, mode="${mode}")
-ModelCheckpointCfg = builds(
-    ModelCheckpoint,
-    dirpath=os.path.join(output_dir, "checkpoints"),
-    monitor="${monitor}",
-    save_last=True,
-    mode="${mode}",
-    auto_insert_metric_name=False,
-)
